backend/requirements.txt
django>=1.7,<1.8
django-grappelli==2.6.4
django-filebrowser==3.5.7
djangorestframework==3.1.1
Markdown==2.6.2 # for API docs
django-filter==0.9.2
pylibmc==1.4.2
sorl-thumbnail==12.2
django_smtp_ssl==1.0
raven==5.3.1
Pillow==2.8.1
django-custom-user==0.5
djoser==0.3.0